@@ -1,5 +1,5 @@
-class AddForeignKeyToClients < ActiveRecord::Migration
-  def change  
+class AddForeignKeyToClients < ActiveRecord::Migration[4.2]
+  def change
     add_reference :clients, :lizards, index: true, foreign_key: true
   end
 end
