@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -11,59 +10,66 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20171117033958) do
+ActiveRecord::Schema.define(version: 20171121115957) do
 
-  create_table "clients", force: :cascade do |t|
-    t.string   "first_name"
-    t.string   "last_name"
-    t.integer  "phone_number"
-    t.string   "address"
-    t.datetime "created_at",   null: false
-    t.datetime "updated_at",   null: false
-    t.integer  "lizards_id"
+  # These are extensions that must be enabled in order to support this database
+  enable_extension "plpgsql"
+
+  create_table "clients", id: :serial, force: :cascade do |t|
+    t.string "first_name"
+    t.string "last_name"
+    t.integer "phone_number"
+    t.string "address"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.integer "lizards_id"
+    t.index ["lizards_id"], name: "index_clients_on_lizards_id"
   end
 
-  add_index "clients", ["lizards_id"], name: "index_clients_on_lizards_id"
-
-  create_table "lizards", force: :cascade do |t|
-    t.string   "petname"
-    t.integer  "weight"
-    t.integer  "length"
-    t.date     "last_seen"
-    t.datetime "created_at",    null: false
-    t.datetime "updated_at",    null: false
-    t.integer  "social_score"
-    t.integer  "diet_score"
-    t.integer  "fitness_score"
-    t.integer  "mental_score"
+  create_table "lizards", id: :serial, force: :cascade do |t|
+    t.string "petname"
+    t.integer "weight"
+    t.integer "length"
+    t.date "last_seen"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.integer "social_score"
+    t.integer "diet_score"
+    t.integer "fitness_score"
+    t.integer "mental_score"
   end
 
-  create_table "users", force: :cascade do |t|
-    t.string   "email",                  default: "", null: false
-    t.string   "encrypted_password",     default: "", null: false
-    t.string   "reset_password_token"
+  create_table "mpayments", force: :cascade do |t|
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+  end
+
+  create_table "users", id: :serial, force: :cascade do |t|
+    t.string "email", default: "", null: false
+    t.string "encrypted_password", default: "", null: false
+    t.string "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",          default: 0,  null: false
+    t.integer "sign_in_count", default: 0, null: false
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
-    t.string   "current_sign_in_ip"
-    t.string   "last_sign_in_ip"
-    t.datetime "created_at",                          null: false
-    t.datetime "updated_at",                          null: false
+    t.string "current_sign_in_ip"
+    t.string "last_sign_in_ip"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["email"], name: "index_users_on_email", unique: true
+    t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
   end
 
-  add_index "users", ["email"], name: "index_users_on_email", unique: true
-  add_index "users", ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
-
-  create_table "veterinarians", force: :cascade do |t|
-    t.string   "first_name"
-    t.string   "last_name"
-    t.integer  "years_in_practice"
-    t.string   "vet_university"
-    t.string   "avatar"
-    t.datetime "created_at",        null: false
-    t.datetime "updated_at",        null: false
+  create_table "veterinarians", id: :serial, force: :cascade do |t|
+    t.string "first_name"
+    t.string "last_name"
+    t.integer "years_in_practice"
+    t.string "vet_university"
+    t.string "avatar"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
+  add_foreign_key "clients", "lizards", column: "lizards_id"
 end
